+++ src/test-integration/resources/jsSearch/level1/level2/level2.js
@@ -1,3 +1,5 @@
 var z = 0;
-if (z > 0)
-    z--;
+function fn(z) {
+    if (z > 0)
+        z--;
+}