+++ src/test-integration/resources/jsSearch/root.js
@@ -1,5 +1,5 @@
 var x = 0;
 if (x > 0)
     x--;
-x++;
-x++;
+function fn(x){return x+1;}
+x = fn(x);